refactor(checkout): CHECKOUT-2951 Convert `quote` module into TypeScript
